Support many occurrences of a tag, for certain tags
Some patches, especially when coming from emails, may contain multiple
occurrences of "References:" and "From:" tags.

get/extract operate on the last occurrence of a tag
add places the new tag after already present ones and, as before, tries to put
it in a sensible place. Add support is not complete (does not support certain
types of tags) but has worked well enough for now.