Process attribution lines individually, instead of as a block
Because of commits like linux.git 51151a1, it is difficult to strictly
delineate attribution "blocks" (a series of lines). This patch changes the
code to no longer extract and replace an attribution block. Instead, it only
inserts or appends new attributions lines. This only affects cases where the
patch being cleaned is missing attribution lines compared to the original
upstream commit message.