Introduce quilt-mode
These tools are similar in spirit to backport-mode but they are meant to
work with kernel-source.git's quilt tree instead of kernel.git's git tree.

Interleaving new patches while working with kernel.git proved to be very
laborious and slow. While git is certainly a more powerful tool than quilt,
quilt is faster when rebasing hundreds of patches and, more significantly,
there is no need for complicated manipulations of kernel.git to simulate
inserting patches in the middle of the series.

! Note that these tools are meant to work with a modified version of quilt
that uses series.conf directly.

Obviously, some constants are hard coded for SLE12-SP3 in qgoto.py.

More tools to come...