sequence-insert: Helper to determine sequence_patch.sh argument
When we want to backport a specific commit at its right position in the
sorted sub-series, it is most efficient to use sequence_patch.sh to expand
the tree up to the patch just before where the new commit will be added.
sequence-insert.py prints out which patch that is.